added authorizations to all controller's actions

+++ app/controllers/categorizations_controller.rb
@@ -5,18 +5,20 @@ class CategorizationsController < ApplicationController
 
 
   def index
+    authorize Item,:create?, policy_class:ItemPolicy
     @categorizations=Categorization.get_assigned_categories
   end
 
 
 
   def new
+    authorize Item,:create?, policy_class:ItemPolicy
     @categorization=Categorization.new
 
   end
 
   def create
-
+    authorize Item,:create?, policy_class:ItemPolicy
     item=Item.find_by(title:params[:categorization][:item])
     category=Category.find_by(name:params[:categorization][:category])
     if item && category
@@ -32,6 +34,7 @@ class CategorizationsController < ApplicationController
   end
 
   def destroy
+    authorize Item,:create?, policy_class:ItemPolicy
     session[:return_to] ||= request.referer
     if @categorization.destroy
       redirect_to session.delete(:return_to),:flash => { :success => "Successfully unassigned category to item" }
